Replace old plot functions with PGFPlotsX
The PseudomodesTTEDOPA package isn't meant to contain plotting
utilities, but just the ITensors part of the calculations (plus some
generic utility function). With this in mind, the `groupplot`,
`unifiedplot` ecc. functions will cease to be provided from the package
sooner or later. At the same time, these functions are not really good
pieces of code, and their graphical output isn't pretty either, so this
is a good occasion to get rid of them.
The perfect replacement is PGFPlotsX, which produces by default
higher-quality plots, and whose syntax is shorter (albeit slightly more
complex), so the previous functions can be replaced without lengthening
the scripts too much.

Note that no tests whatsoever were done to ascertain that the code has
been correctly replaced (there are some unrelated errors anyway that
must be corrected first).