Fix missing definition of matrices and `kron`
Now that the definitions of states and operators are gone, those of the
"elementary" matrices of harmonic oscillators and spins, such as Pauli
matrices of ladder operators, are no longer there too.
Since I don't want to export them in PseudomodesTTEDOPA, I added their
definitions file by file when needed. This required also adding the
definition of the tensor product and adding the LinearAlgebra and
QuantumOptics packages in some scripts (they were previously included by
files in the `lib` directory).