Move `const \otimes = kron` inside `let` block
Some script use this statement in order to use the mathematical symbol
as a shorthand for the operator. However, then they were included into
the REPL, the statement was not confined to the `let` block therefore it
ended up "polluting" the REPL. For example, QuantumOptics also defines
that symbol in another way, which caused some MethodErrors unless the
REPL was restarted each time.