Convert closed_spin_chain from TEBD to TDVP
Thanks to the ITensorTDVP library I was able to use TDVP to simulate the
system in this first example. I had to cut some parts away from the
script, but they will be restored whenever the opportunity arises.
For example, since we're using 1-site TDVP, there's no point in
computing the bond dimensions of the state.